% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_environ_packages.R
\name{add_environ_packages}
\alias{add_environ_packages}
\title{Add packages from current environment to dockerfile}
\usage{
add_environ_packages(dockerfile)
}
\arguments{
\item{dockerfile, }{a \code{dockerfile} object}
}
\value{
a dockerfile object
}
\description{
Add packages from current environment to dockerfile
}
\examples{
dockerfile() \%>\%
  get_current_package_dependencies()

}
