% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dockerfile.R
\name{basic_dockerfile}
\alias{basic_dockerfile}
\title{Pre-formats and opens a new dockerfile}
\usage{
basic_dockerfile(open = TRUE)
}
\arguments{
\item{open, }{whether or not to open the basic dockerfile after its created}
}
\value{
nothing
}
\description{
Pre-formats and opens a new dockerfile
}
\examples{
 basic_dockerfile()

}
