% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build an image from a Dockerfile}
\usage{
build(path = ".", tag = NULL)
}
\arguments{
\item{path}{A file path to the Dockerfile, defaults to the current working
directory}

\item{tag}{A tag to add to the final image}
}
\value{
the built Docker image object
}
\description{
Build an image from a Dockerfile
}
\examples{
\dontrun{
build(path = "Dockerfiles/helloworld")
}
}
