% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_image.R
\name{build_image}
\alias{build_image}
\title{Build an image from a Dockerfile}
\usage{
build_image(path = here::here(), tag = NULL)
}
\arguments{
\item{path}{A file path to the Dockerfile, defaults to the current working
directory}

\item{tag}{A tag to add to the final image}
}
\value{
the built Docker image object
}
\description{
Build an image from a Dockerfile
}
\examples{
build_image(path = "Dockerfiles/helloworld")
}
