% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_from.R
\name{check_for_from}
\alias{check_for_from}
\title{check_for_from}
\usage{
check_for_from(dockerfile)
}
\arguments{
\item{dockerfile}{object from dockerfile()}
}
\value{
\code{from_check}list of true/false checks
}
\description{
check_for_from
}
\examples{
dockerfile() \%>\% set_from(., "debian:stretch") \%>\% check_for_from()

}
