% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_current_dockerfile.R
\name{create_current_dockerfile}
\alias{create_current_dockerfile}
\title{Creates a dockerfile that mirrors the current environment}
\usage{
create_current_dockerfile(file = "Dockerfile")
}
\arguments{
\item{file}{\code{String} output dockerfile location and name, defaults to "Dockerfile"}
}
\value{
\code{String} output dockerfile location and name
}
\description{
Creates a dockerfile that mirrors the current environment
}
\examples{
\dontrun{
create_current_dockerfile()
}
}
