% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_build.R
\name{docker_build}
\alias{docker_build}
\title{Build an image from a Dockerfile}
\usage{
docker_build(path = here::here(), tag = NULL)
}
\arguments{
\item{path}{A file path to the Dockerfile, defaults to the current working
directory}

\item{tag}{A tag to add to the final image}
}
\value{
the built Docker image object
}
\description{
Build an image from a Dockerfile
}
\examples{
docker_build(path = "Dockerfiles/helloworld")
}
