% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{docker_image_info}
\alias{docker_image_info}
\title{Get info about a Docker iamge}
\usage{
docker_image_info(img)
}
\arguments{
\item{img}{a Docker image object}
}
\value{
list of the image attributes
}
\description{
Get info about a Docker iamge
}
