% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_images.R
\name{docker_images}
\alias{docker_images}
\title{List Docker images}
\usage{
docker_images(name = NULL, all = NULL, filters = NULL)
}
\arguments{
\item{name}{Only show images belonging to the repository name}

\item{all}{Show intermediate image layers. By default, these are filtered out}

\item{filters}{Filters to be processed on the image list.}
}
\value{
list of images
}
\description{
List Docker images
}
\examples{
docker_images()
}
