% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_list.R
\name{docker_list}
\alias{docker_list}
\title{List Docker images}
\usage{
docker_list(name = NULL, all = NULL, filters = NULL)
}
\arguments{
\item{name}{Only show images belonging to the repository name}

\item{all}{Show intermediate image layers. By default, these are filtered out}

\item{filters}{Filters to be processed on the image list.}
}
\value{
list of images
}
\description{
List Docker images
}
\examples{
docker_list()
}
