% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_pull.R
\name{docker_pull}
\alias{docker_pull}
\title{Pull a Docker image from a repository}
\usage{
docker_pull(repository, tag = "latest")
}
\arguments{
\item{repository}{the repository to pull}

\item{tag}{the tag to pull (optional)}
}
\description{
Pull a Docker image from a repository
}
\examples{
docker_pull("rocker/tidyverse")
}
