% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker_run}
\alias{docker_run}
\title{Run a Docker container}
\usage{
docker_run(image, command = NULL)
}
\arguments{
\item{image}{the name of the Docker container to run}

\item{command}{the command to run in the container (optional)}
}
\value{
the container output
}
\description{
Run a Docker container
}
\examples{
docker_run("alpine", "echo hello world")
}
