% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_tag.R
\name{docker_tag}
\alias{docker_tag}
\title{Tag a Docker image}
\usage{
docker_tag(source, target)
}
\arguments{
\item{source}{the name of a source image}

\item{target}{the tag name}
}
\value{
the tagged image
}
\description{
Create a tag "target" that refers to "source" image
}
\examples{
\dontrun{
docker_tag("rocker/tidyverse", "v1.2.3.9000")
}
}
