% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{dockerfile}
\alias{dockerfile}
\title{dockerfile creates a basic, empty docker object}
\usage{
dockerfile()
}
\value{
\code{dockerfile}, a specification of the contents of the dockerfile
}
\description{
dockerfile creates a basic, empty docker object
}
\examples{
dockerfile()

}
