% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_package_dependencies.R
\name{get_current_package_dependencies}
\alias{get_current_package_dependencies}
\title{Gets packages that are loaded in the current envrionment, their dependencies, and the versions for all}
\usage{
get_current_package_dependencies()
}
\value{
\code{tibble} of packages and their required version
}
\description{
Gets packages that are loaded in the current envrionment, their dependencies, and the versions for all
}
\examples{
get_current_package_dependencies()
}
