% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_blank.R
\name{insert_blank}
\alias{insert_blank}
\title{Adds a blank line at the end of the current dockerfile}
\usage{
insert_blank(dockerfile)
}
\arguments{
\item{dockerfile}{}
}
\value{
\code{dockerfile}, a specification of the contents of the dockerfile
}
\description{
Adds a blank line at the end of the current dockerfile
}
\examples{
dockerfile() \%>\%
  insert_blank() \%>\%
  write_dockerfile()
}
