% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_docker.R
\name{install_docker}
\alias{install_docker}
\title{Install python docker package}
\usage{
install_docker(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method, passed to py_install}

\item{conda}{Path to conda, passed to py_install}
}
\description{
Install python docker package
}
\examples{
\dontrun{
install_docker()
}
}
