% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull a Docker image from a repository}
\usage{
pull(repository, tag = "latest")
}
\arguments{
\item{repository}{the repository to pull}

\item{tag}{the tag to pull (optional)}
}
\description{
Pull a Docker image from a repository
}
\examples{
pull("rocker/tidyverse")
}
