% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{set_env}
\alias{set_env}
\title{Title}
\usage{
set_env(dockerfile, env_vars)
}
\arguments{
\item{dockerfile}{an object of class \code{dockerfile}}

\item{env_vars}{A named vector of environment variables and their values}
}
\value{
\code{dockerfile}, a specification of the contents of the dockerfile
}
\description{
Title
}
\examples{
dockerfile() \%>\%
  set_env(c("S6_VERSION" = "${S6_VERSION:-v1.21.7.0}",
            "S6_BEHAVIOUR_IF_STAGE_2_FAILS", "2")) \%>\%
  write_dockerfile()

}
