% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_expose.R
\name{set_expose}
\alias{set_expose}
\title{Set which ports the docker image can expose, and what protocols to use}
\usage{
set_expose(dockerfile, port, protocol = c("tcp"))
}
\arguments{
\item{dockerfile, }{an object of class \code{dockerfile}}

\item{port}{the port to expose}

\item{protocol}{charater vector of protocols to use. Defalts to 'tcp', but can take on 'tcp' or 'udp'}
}
\value{
\code{dockerfile}
}
\description{
Set which ports the docker image can expose, and what protocols to use
}
\examples{
dockerfile() \%>\%
set_expose(8080)
}
