% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_from.R
\name{set_from}
\alias{set_from}
\title{set_from}
\usage{
set_from(dockerfile, image)
}
\arguments{
\item{dockerfile, }{an object of class \code{dockerfile}}

\item{image}{an object of class \code{character}}
}
\value{
an object of class \code{dockerfile}
}
\description{
set_from
}
\examples{
dockerfile() \%>\%
 set_from("debian:stretch")

}
