% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_label.R
\name{set_label}
\alias{set_label}
\title{Add label metadata to an image}
\usage{
set_label(dockerfile, label_vars)
}
\arguments{
\item{dockerfile}{a specification of the contents of the dockerfile}

\item{label_vars}{A named vector of labels and their values}
}
\value{
\code{dockerfile}, a specification of the contents of the dockerfile
}
\description{
Add label metadata to an image
}
\examples{
dockerfile() \%>\%
set_label(c("label_1"="value_1", "label_2"="value_2"))

}
