% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_run.R
\name{set_run}
\alias{set_run}
\title{Set a command to run in the image}
\usage{
set_run(dockerfile, command, blank = TRUE)
}
\arguments{
\item{dockerfile}{a specification of the contents of the dockerfile}

\item{command}{a command to run in the shell}

\item{blank}{boolean indicating whether or not to add a trailing space}
}
\value{
\code{dockerfile}
}
\description{
Set a command to run in the image
}
\examples{
dockerfile() \%>\%
  set_run("-R install.packages('ggplot2')")
}
