% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_current_dockerfile.R
\name{create_current_dockerfile}
\alias{create_current_dockerfile}
\title{Creates a dockerfile that mirrors the current environment}
\usage{
create_current_dockerfile(file = "Dockerfile")
}
\arguments{
\item{`String`}{output dockerfile location and name, defaults to "Dockerfile"}
}
\description{
Creates a dockerfile that mirrors the current environment
}
\examples{
create_current_dockerfile()
}
