% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{docker}
\alias{docker}
\title{Docker module object}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
docker
}
\description{
Uses the reticulate framework to access the Docker SDK.
}
\keyword{datasets}
