% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cmd.R
\name{set_cmd}
\alias{set_cmd}
\title{Set a command to executed when running the image}
\usage{
set_cmd(dockerfile, command, blank = TRUE)
}
\arguments{
\item{dockerfile}{a dockerfile object}

\item{command}{a command to be passed to the dockerfile}

\item{blank}{whether to include a blank line at the end of file}
}
\value{
a dockerfile object
}
\description{
Set a command to executed when running the image
}
