% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dockerfile.R
\name{write_dockerfile}
\alias{write_dockerfile}
\title{Write a dockerfile from a dockerfile object}
\usage{
write_dockerfile(dockerfile, file = "Dockerfile")
}
\arguments{
\item{dockerfile, }{an object of class \code{dockerfile}}
}
\value{
an object of class \code{dockerfile}
}
\description{
Write a dockerfile from a dockerfile object
}
\examples{
dockerfile() \%>\%
  insert_blank() \%>\%
  write_dockerfile()

}
